chore: prepare release v0.1.1
Prepare for v0.1.1 release with changelog generation and pre-merge
validation.

Changes:
- Generated CHANGELOG.md using git-cliff for v0.1.1
- Bumped version to 0.1.1 in Cargo.toml
- Fixed clippy warning: unnecessary cast in portal_publishing.rs
- Updated bytes dependency: 1.11.0 → 1.11.1 (security fix RUSTSEC-2026-0007)
- Applied cargo fmt to all files

Pre-merge validation complete:
✅ just check-all - All checks passed (clippy, fmt, tests)
✅ just audit - No vulnerabilities (1 allowed unmaintained warning)
✅ just check-features - All feature combinations build successfully

Release highlights (v0.1.0 → v0.1.1):
- Major geometry module consolidation (ESRI integration complete)
- Service definition strong typing (6 phases complete)
- Portal addToDefinition implementation
- ESRI error response checking (prevents silent failures)
- New examples: geoprocessing, elevation, vector tiles
- Example assertion pattern for validation
- Comprehensive gap analysis (35% coverage documented)

Testing:
- 253 doctests passing
- 33 unit tests passing
- 14 working examples (11 enterprise, 3 public)

+++ examples/enterprise/portal_publishing.rs
@@ -87,9 +87,9 @@
 
 use anyhow::{Context, Result};
 use arcgis::{
-    AddItemParams, AddToDefinitionParams, ApiKeyAuth, ApiKeyTier, ArcGISClient, CreateServiceParams,
-    EditOptions, Feature, FeatureServiceClient, FieldDefinitionBuilder, FieldType,
-    GeometryTypeDefinition, LayerDefinitionBuilder, LayerId, NoAuth, PortalClient,
+    AddItemParams, AddToDefinitionParams, ApiKeyAuth, ApiKeyTier, ArcGISClient,
+    CreateServiceParams, EditOptions, Feature, FeatureServiceClient, FieldDefinitionBuilder,
+    FieldType, GeometryTypeDefinition, LayerDefinitionBuilder, LayerId, NoAuth, PortalClient,
     SharingParameters,
 };
 use std::collections::HashMap;
@@ -202,7 +202,9 @@ async fn demonstrate_workflow_a_direct_service(
     tracing::info!("\n========================================");
     tracing::info!("=== WORKFLOW A: Direct Service Creation ===");
     tracing::info!("========================================");
-    tracing::info!("Query → Convert → Create Service → Add Definition → Add Features → Share → Verify → Cleanup");
+    tracing::info!(
+        "Query → Convert → Create Service → Add Definition → Add Features → Share → Verify → Cleanup"
+    );
     tracing::info!("");
 
     // ========================================================================
@@ -342,7 +344,9 @@ async fn demonstrate_workflow_a_direct_service(
     tracing::info!("");
 
     let add_def_params = AddToDefinitionParams::new().with_layers(vec![layer]);
-    let add_def_result = portal.add_to_definition(&service_item_id, add_def_params).await?;
+    let add_def_result = portal
+        .add_to_definition(&service_item_id, add_def_params)
+        .await?;
 
     // Assertion: Operation must succeed
     assert!(
@@ -421,7 +425,7 @@ async fn demonstrate_workflow_a_direct_service(
 
     // Assertion: Number of successes should match features we tried to add
     assert_eq!(
-        edit_result.success_count() as usize,
+        edit_result.success_count(),
         features_to_add.len(),
         "Expected {} successes, got {}",
         features_to_add.len(),
